Module Irmin_containers.Blob_log

The implementation of log in which it is maintained as a single unit, or blob. Hence, two versions of the log cannot share their common predecessor. The type of values to be stored as well as a method to obtain timestamps are provided by the user.

Merging does the following: the newer entries from each branch, with respect to the least common ancestor, are taken, merged and then appended in front of the LCA.

module type S = sig ... end

Signature of Blob_log

module Make (Backend : sig ... end) (T : Time.S) (V : Irmin.Type.S) : S with type value = V.t and type Store.branch = string and type Store.key = string list and type Store.step = string

Make returns a mergeable blob log using the backend and other parameters as specified by the user.

module FS (V : Irmin.Type.S) : S with type value = V.t and type Store.branch = string and type Store.key = string list and type Store.step = string

Blob log instantiated using the FS backend provided by Irmin_unix and the timestamp method Time.Unix

module Mem (V : Irmin.Type.S) : S with type value = V.t and type Store.branch = string and type Store.key = string list and type Store.step = string

Blob log instantiated using the in-memory backend provided by Irmin_mem and the timestamp method Time.Unix