Module Irmin_containers.Counter

The implementation of an int64 counter. This module supports operations to increment, decrement and read the value of the counter.

Merge semantics is as follows: if old is the value of the LCA and v1 and v2 are the current values, then the merged value is v1 + v2 - old.

module type S = sig ... end

Counter signature

module Make : functor (Backend : Irmin.KV_maker) -> S

Make returns a mergeable counter using the backend and other parameters as specified by the user.

module FS : S

Counter instantiated using the FS backend provided by Irmin_unix

module Mem : S

Counter instantiated using the in-memory backend provided by Irmin_mem