Module Irmin_containers.Lww_register

The implementation of last-write-wins register. The value to be stored in the register and the timestamp method are provided by the user.

Merge semantics: The value with the largest timestamp is chosen. If two values have the same timestamp, then the larger value is selected based on the compare specified by the user.

module type S = sig ... end

Signature of Lww_register

module Make (Backend : Irmin.KV_maker) (T : Time.S) (V : Irmin.Type.S) : S with type value = V.t

Make returns a mergeable last-write-wins register using the backend and other parameters as specified by the user.

module FS (V : Irmin.Type.S) : S with type value = V.t

LWW register instantiated using the FS backend provided by Irmin_unix and the timestamp method Time.Unix

module Mem (V : Irmin.Type.S) : S with type value = V.t

LWW register instantiated using the in-memory backend provided by Irmin_mem and the timestamp method Time.Unix