Module Irmin_containers__.Blob_log

module type S = sig ... end

Signature of Blob_log

module Make : functor (Backend : Irmin_containers__.Stores.Store_maker) -> functor (T : Irmin_containers.Time.S) -> functor (V : Irmin.Type.S) -> S with type value = V.t and type Store.branch = string and type Store.key = string list and type Store.step = string

Make returns a mergeable blob log using the backend and other parameters as specified by the user.

module FS : functor (V : Irmin.Type.S) -> S with type value = V.t and type Store.branch = string and type Store.key = string list and type Store.step = string

Blob log instantiated using the FS backend provided by Irmin_unix and the timestamp method Time.Unix

module Mem : functor (V : Irmin.Type.S) -> S with type value = V.t and type Store.branch = string and type Store.key = string list and type Store.step = string

Blob log instantiated using the in-memory backend provided by Irmin_mem and the timestamp method Time.Unix