Module Linked_log.FS

Linked log instantiated using the FS backend provided by Irmin_unix, timestamp method Time.Unix and hash Irmin.Hash.SHA1

Parameters

C : Irmin_containers__.Stores.Cas_maker

V : Irmin.Type.S

Signature

module Store : Irmin.S

Store for the log. All store related operations like branching, cloning, merging, etc are done through this module.

type value

Type of log entry

val append : path:Store.key -> Store.t -> value -> unit Lwt.t

Append an entry to the log

val read_all : path:Store.key -> Store.t -> value list Lwt.t

Read the entire log

type cursor

Type of cursor. Cursor is like a marker from which a certain number of entries can be read

val get_cursor : path:Store.key -> Store.t -> cursor Lwt.t

Create a new cursor over the log entires at the given path

val read : num_items:int -> cursor -> (value list * cursor) Lwt.t

Read at most num_items entries from the cursor. If the number specified is greater than the number of log entries from the cursor, the log is read till the end. If the input cursor has already reached the end, then an empty list is returned