Module Irmin_containers__Counter.Make

Make returns a mergeable counter using the backend and other parameters as specified by the user.

Parameters

Backend : Irmin.KV_maker

Signature

module Store : Irmin.KV

Content store of counter. All store related operations like branching, cloning, merging, etc are done through this module.

val inc : ?⁠by:int64 -> ?⁠info:Store.Info.f -> path:Store.path -> Store.t -> unit Lwt.t

Increment the counter by the amount specified using by. If no value is specified, then by is assigned the value 1L.

val dec : ?⁠by:int64 -> ?⁠info:Store.Info.f -> path:Store.path -> Store.t -> unit Lwt.t

Decrement the counter by the amount specified using by. If no value is specified, then by is assigned the value 1L.

val read : path:Store.path -> Store.t -> int64 Lwt.t

Read the value of the counter