Module Irmin_containers__Lww_register

The implementation of last-write-wins register. The value to be stored in the register and the timestamp method are provided by the user.

Merge semantics: The value with the largest timestamp is chosen. If two values have the same timestamp, then the larger value is selected based on the compare specified by the user.