Module Irmin_containers__Lww_register.Make

Make returns a mergeable last-write-wins register using the backend and other parameters as specified by the user.

Parameters

Backend : Irmin.KV_maker

T : Irmin_containers.Time.S

V : Irmin.Type.S

Signature

module Store : Irmin.KV

Content store of the register. All store related operations like branching, cloning, merging, etc are done through this module.

type value = V.t

Type of values stored in the register

val read : path:Store.path -> Store.t -> value option Lwt.t

Reads the value from the register. Returns None if no value is written

val write : ?⁠info:Store.Info.f -> path:Store.path -> Store.t -> value -> unit Lwt.t

Writes the provided value to the register