Module Irmin_containers__Stores

module type Store_maker = functor (C : Irmin.Contents.S) -> Irmin.S with type contents = C.t and type branch = string and type key = string list and type step = string

The signature for the backend input to the data structures. The Irmin stores of the data structures are constructed using modules of this type

module type Cas_maker = sig ... end