Module Irmin_containers

Irmin_containers is a collection of simple, ready-to-use mergeable data structures. Each data structure works with an arbitrary Irmin backend and is customisable in a variety of ways.

Additionally, Irmin_containers supplies instantiations of each of these data structures with two backends:

the in-memory backend provided by Irmin_mem

the FS backend provided by Irmin_unix.

Data structures

module Counter : sig ... end

module Lww_register : sig ... end

module Blob_log : sig ... end

module Linked_log : sig ... end

Auxiliary signatures and modules

module type Store_maker = sig ... end

Store_maker is the signature for the backend input to the data structures. The Irmin stores of the data structures are constructed using modules of this type

module type Cas_maker = sig ... end

Cas_maker is the signature for the store which will be used to maintain linked data structures. The elements are hashed into this store and the hash value is used to construct the linkages.

module Time : sig ... end