Module Blob_log.Make

Make returns a mergeable blob log using the backend and other parameters as specified by the user.

Parameters

module Backend : sig ... end

module T : Time.S

module V : Irmin.Type.S

Signature

module Store : Irmin.S with type branch = string with type key = string list with type step = string

Store for the log. All store related operations like branching, cloning, merging, etc are done through this module.

type value = V.t

Type of log entry

val append : path:Store.key -> Store.t -> value -> unit Lwt.t

Append an entry to the log

val read_all : path:Store.key -> Store.t -> value list Lwt.t

Read the entire log