Module Irmin_containers.Lww_register

The implementation of last-write-wins register. The value to be stored in the register and the timestamp method are provided by the user.

Merge semantics: The value with the largest timestamp is chosen. If two values have the same timestamp, then the larger value is selected based on the compare specified by the user.

module type S = sig ... end

Signature of Lww_register

module Make : functor (Backend : Irmin_containers__.Stores.Store_maker) -> functor (T : Time.S) -> functor (V : Irmin.Type.S) -> S with type value = V.t and type Store.branch = string and type Store.key = string list and type Store.step = string

Make returns a mergeable last-write-wins register using the backend and other parameters as specified by the user.

module FS : functor (V : Irmin.Type.S) -> S with type value = V.t and type Store.branch = string and type Store.key = string list and type Store.step = string

LWW register instantiated using the FS backend provided by Irmin_unix and the timestamp method Time.Unix

module Mem : functor (V : Irmin.Type.S) -> S with type value = V.t and type Store.branch = string and type Store.key = string list and type Store.step = string

LWW register instantiated using the in-memory backend provided by Irmin_mem and the timestamp method Time.Unix